[REFACTOR][BE][석범] 서비스 위치 변경

+++ BE/peaktime/src/main/java/com/dinnertime/peaktime/domain/hiking/repository/HikingRepositoryCustom.java
@@ -19,5 +19,5 @@ public interface HikingRepositoryCustom {
 
     Integer getPreferTimeByUserId(Long userId);
 
-    Integer getTotalBlockedCount(Long findUserId);
+    Long getTotalBlockedCount(Long findUserId);
 }
